fix(TextArea): broken animation when scrollHeight is 0
