+++ packages/ui-components/src/components/TextArea/__tests__/utilities.test.ts
@@ -1,0 +1,7 @@
+import { adjustLabelAndHelperText } from "../utilities";
+
+describe("adjustLabelAndHelperText", () => {
+	it("should", () => {
+		expect(true).toBe(true);
+	});
+});
